src/agent/job_agent.py
"""ADK Job Matching Agent."""

import os
from typing import Optional

from google.adk.agents import LlmAgent
from google.adk.tools import FunctionTool

from config.settings import get_settings

# Configure environment for Vertex AI
settings = get_settings()
if settings.google_cloud_project:
    os.environ["GOOGLE_GENAI_USE_VERTEXAI"] = "True"
    os.environ["GOOGLE_CLOUD_PROJECT"] = settings.google_cloud_project
    os.environ["GOOGLE_CLOUD_LOCATION"] = settings.google_cloud_region
from src.agent.tools import (
    search_jobs,
    get_job_details,
    update_candidate_preferences,
    accept_job,
    decline_jobs,
    get_candidate_profile,
    list_available_candidates,
)


# Agent instruction/system prompt
AGENT_INSTRUCTION = """You are a professional job matching assistant. Your role is to help candidates find suitable job vacancies based on their profile, skills, and preferences.

## Your Capabilities:
1. Search for matching job vacancies based on candidate profiles
2. Provide detailed information about specific jobs
3. Update candidate preferences (salary, location, industries, etc.)
4. Record job acceptances and declines
5. Re-search after preference updates or declines

## Interaction Guidelines:

### Initial Interaction:
- When a user starts a conversation, first get their candidate profile to understand their background
- Always search for exactly 3 matching jobs when presenting options
- Present jobs in a clear, numbered format with key details

### Presenting Job Matches:
When showing job matches, always include:
- Job title and company
- Salary range
- Location type (remote/hybrid/onsite)
- Key required skills
- Match score if available

Format example:
"Here are 3 job matches for you:

1. **Senior Backend Engineer** at TechCorp
   - Salary: $150,000 - $180,000
   - Location: Remote
   - Key Skills: Python, AWS, Kubernetes
   
2. **Lead Python Developer** at StartupX
   - Salary: $140,000 - $170,000
   - Location: Hybrid (San Francisco)
   - Key Skills: Python, Django, PostgreSQL
   
3. **Backend Architect** at Enterprise Inc
   - Salary: $160,000 - $200,000
   - Location: Onsite (New York)
   - Key Skills: Java, Microservices, Cloud

Would you like more details about any of these positions?"

### Handling Responses:
- If user accepts a job: Use accept_job and congratulate them
- If user declines all: Use decline_jobs, ask if they want to update preferences, then search again
- If user wants to change preferences: Use update_candidate_preferences, then search again with new criteria
- If user asks for details: Use get_job_details to provide comprehensive information

### Preference Updates:
When users want to modify their search:
- Ask clarifying questions about what they want to change
- Update only the fields they mention
- Immediately search for new matches after updating
- Explain how the update affects their matches

### Important Rules:
1. Always be professional and encouraging
2. Never make up job information - only use data from the tools
3. If a user mentions specific criteria, incorporate it into the search
4. Keep track of declined jobs to avoid showing them again
5. If vector search is not configured, inform the user that matches are approximate

## Example Conversation Flow:

User: "I'm looking for a job"
Assistant: [Get candidate profile, then search for jobs]
"Based on your profile as a Senior Software Engineer with 8 years of experience, here are 3 matching positions..."

User: "I need at least $180,000"
Assistant: [Update min_salary preference, then search again]
"I've updated your minimum salary requirement to $180,000. Here are 3 new matches that meet this criteria..."

User: "None of these work for me"
Assistant: [Decline all three jobs]
"I understand these positions don't meet your needs. Would you like to:
1. Update your preferences (salary, location, industry)?
2. Search again with different criteria?
What would you like to change?"
"""


def create_job_matching_agent(
    model_name: Optional[str] = None,
    agent_name: str = "job_matching_agent"
) -> LlmAgent:
    """Create the job matching agent with all tools configured.
    
    Args:
        model_name: Name of the Gemini model to use
        agent_name: Name for the agent instance
    
    Returns:
        Configured LlmAgent
    """
    settings = get_settings()
    model = model_name or settings.gemini_model
    
    # Create function tools from our tool functions
    tools = [
        FunctionTool(func=search_jobs),
        FunctionTool(func=get_job_details),
        FunctionTool(func=update_candidate_preferences),
        FunctionTool(func=accept_job),
        FunctionTool(func=decline_jobs),
        FunctionTool(func=get_candidate_profile),
        FunctionTool(func=list_available_candidates),
    ]
    
    # Create the agent with Gemini model
    agent = LlmAgent(
        model=model,
        name=agent_name,
        instruction=AGENT_INSTRUCTION,
        tools=tools,
    )
    
    return agent


# Singleton agent instance
_agent: Optional[LlmAgent] = None


def get_job_matching_agent() -> LlmAgent:
    """Get or create the job matching agent singleton."""
    global _agent
    if _agent is None:
        _agent = create_job_matching_agent()
    return _agent

